requirements.txt
# Core dependencies
crewai
crewai-tools
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-openai
langchain-community
pypdf
openai
pandas
numpy<2.0.0
sqlalchemy
redis
celery